Use deltas for some output fields
Several types of records have very impressive locality. When a call is
made to a function in a given file, the next call will often be to
another function in the same file, or the same package. When unwinding
stacks, the next instruction pointer will often be in the same shared
library as the last one, and so near it in memory.

Because of this, we can represent many types of records much more
succinctly as an delta from the last value of that type that we saw.

+++ src/memray/_memray/records.h
@@ -203,7 +203,7 @@ struct FramePop
 struct UnresolvedNativeFrame
 {
     uintptr_t ip;
-    uint32_t index;
+    size_t index;
 };
 
 struct ContextSwitch
